Check auth for pricing
